# slides/se332/openapi.yaml
...
openapi: 3.0.0
info:
  title: UIT Course & Student Manager API
  version: 1.0.0
  description: Comprehensive CRUD API powered by Redis Stack (RedisJSON + RediSearch)
...
paths:
...
  /courses:
    post:
      summary: Create a new course (JSON.SET)
      tags: [course]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourseCreate'
      responses:
        201:
          description: Course created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Course'
    get:
      summary: List all courses with pagination
      tags: [course]
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        200:
          description: List of courses
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedCourses'
  /courses/{id}:
    get:
      summary: Get course by ID (JSON.GET)
      tags: [course]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: Course found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Course'
        404:
          description: Course not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      summary: Partially update course (JSON.SET on paths)
      tags: [course]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourseUpdate'
      responses:
        200:
          description: Course updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Course'
        404:
          description: Course not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      summary: Replace entire course record
      tags: [course]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourseCreate'
      responses:
        200:
          description: Course replaced
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Course'
        404:
          description: Course not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Delete course by ID
      tags: [course]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        204:
          description: Course deleted
        404:
          description: Course not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /courses/search:
    get:
      summary: Search courses (FT.SEARCH with filters)
      tags: [course]
      parameters:
        - name: query
          in: query
          schema:
            type: string
          description: Full-text search (name, lecturer)
        - name: lecturer
          in: query
          schema:
            type: string
          description: Filter by lecturer name
        - name: hasCapacity
          in: query
          schema:
            type: string
            enum: ['true', full]
          description: Filter courses with available seats or courses that are full
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [name, lecturer, enrolled, capacity]
            default: name
        - name: order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: asc
      responses:
        200:
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseSearchResult'
  /courses/{courseId}/students:
    post:
      summary: Enroll student in course (JSON.SET + XADD)
      description: Updates Course/Student JSON documents and appends an enrollment stream event
      tags: [course]
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                studentId:
                  type: string
      responses:
        201:
          description: Student enrolled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnrollmentResult'
    get:
      summary: Get all enrolled students in course
      tags: [course]
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: List of enrolled students
          content:
            application/json:
              schema:
                type: object
                properties:
                  courseId:
                    type: string
                    example: SE332.Q21
                  students:
                    type: array
                    items:
                      $ref: '#/components/schemas/Student'
                  total:
                    type: integer
                    example: 3
  /courses/{courseId}/students/{studentId}:
    delete:
      summary: Unenroll student from course (JSON.SET + XADD)
      description: Updates Course/Student JSON documents and appends an unenrollment stream event
      tags: [course]
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: string
        - name: studentId
          in: path
          required: true
          schema:
            type: string
      responses:
        204:
          description: Student unenrolled
        404:
          description: Course or student not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: UUID
  schemas:
...
    Student:
      type: object
      properties:
        id:
          type: string
          example: 23521476
        username:
          type: string
          example: thiendp
        name:
          type: string
          example: Dang Phu Thien
        cohort:
          type: string
          example: 23
        gpa:
          type: number
          example: 9.2
        enrollments:
          type: array
          items:
            type: string
          example: [SE332.Q21, SE121.Q21]
...
    Course:
      type: object
      properties:
        id:
          type: string
          example: SE332.Q21
        name:
          type: string
          example: Chuyen de CSDL nang cao
        lecturer:
          type: object
          properties:
            name:
              type: string
              example: Tran Thi Hong Yen
        practice:
          type: array
          items:
            type: string
        capacity:
          type: integer
          example: 30
        enrolled:
          type: integer
          example: 3
        students:
          type: array
          items:
            type: string
    CourseCreate:
      type: object
      required:
        - id
        - name
        - lecturer
        - capacity
      properties:
        id:
          type: string
        name:
          type: string
        lecturer:
          type: object
          properties:
            name:
              type: string
        practice:
          type: array
          items:
            type: string
        capacity:
          type: integer
    CourseUpdate:
      type: object
      properties:
        name:
          type: string
        lecturer:
          type: object
          properties:
            name:
              type: string
        practice:
          type: array
          items:
            type: string
        capacity:
          type: integer
    PaginatedCourses:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Course'
        total:
          type: integer
          description: Total number of courses matching the query
          example: 20
        limit:
          type: integer
          example: 10
        offset:
          type: integer
          example: 0
    CourseSearchResult:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Course'
        total:
          type: integer
          description: Total number of matching courses
          example: 3
        query:
          type: string
          description: The search query used
          example: database
    EnrollmentResult:
      type: object
      properties:
        courseId:
          type: string
          example: SE332.Q21
        studentId:
          type: string
          example: 23521476
        enrolledAt:
          type: string
          format: date-time
          example: 2026-03-14T10:00:00Z
        streamId:
          type: string
          description: Redis Stream entry ID from XADD
          example: 1741950000000-0
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: Not found
        message:
          type: string
          example: Student with id 23521476 not found